pyproject.toml
[project]
name = "verifiedsignal"
version = "0.1.0"
description = "A Document Intelligence Platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.30",
    "pydantic-settings>=2.0",
    "sse-starlette>=2.0",
    "arq>=0.26",
    "redis>=4.2",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.2",
    "httpx>=0.27",
    "boto3>=1.35",
    "supabase>=2.3.4",
    "python-jose[cryptography]>=3.3.0",
    "email-validator>=2.3.0",
    "pypdf>=5.0",
    "python-docx>=1.1",
]

[project.scripts]
verifiedsignal = "verifiedsignal.cli:main"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.6",
    "reportlab>=4.0",
]

[tool.pdm]
distribution = true

[tool.pdm.build]
package-dir = "src"

[tool.pdm.scripts]
api = "python -m uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload"
api-prod = "python -m uvicorn app.main:app --host 0.0.0.0 --port 8000"
worker = "python -m arq worker.main.WorkerSettings"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
pythonpath = ["."]
markers = [
    "unit: fast tests with no external services",
    "integration: requires DATABASE_URL and migrations applied to Postgres",
    "e2e: requires Docker CLI (compose file validation)",
    "api: ASGI / HTTP smoke tests (TestClient)",
]
[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests", "app", "worker"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
